src/main/java/cd/go/authorization/guest/executor/RoleConfigValidateRequestExecutor.java
package cd.go.authorization.guest.executor;

import cd.go.authorization.guest.model.RoleConfiguration;
import com.google.gson.Gson;
import com.thoughtworks.go.plugin.api.request.GoPluginApiRequest;
import com.thoughtworks.go.plugin.api.response.DefaultGoPluginApiResponse;
import com.thoughtworks.go.plugin.api.response.GoPluginApiResponse;

import java.util.List;
import java.util.Map;

public class RoleConfigValidateRequestExecutor implements RequestExecutor {
    private static final Gson GSON = new Gson();
    private final GoPluginApiRequest request;

    public RoleConfigValidateRequestExecutor(GoPluginApiRequest request) {
        this.request = request;
    }

    @Override
    public GoPluginApiResponse execute() throws Exception {
        Map<String, String> properties = GSON.fromJson(request.requestBody(), Map.class);

        final List<Map<String, String>> validationResult = RoleConfiguration.validate(properties);
        return DefaultGoPluginApiResponse.success(GSON.toJson(validationResult));
    }
}
